+++ RLSimion/features.h
@@ -8,7 +8,7 @@ struct Feature
 	Feature(){}
 	Feature(unsigned int index, double factor){m_index= index; m_factor= factor;};
 
-	unsigned int m_index;
+	size_t m_index;
 	double m_factor;
 };
 
@@ -20,15 +20,15 @@ enum OverwriteMode{Replace,Add,AllowDuplicates};
 class FeatureList
 {
 	const char *m_name;
-	unsigned int m_numAllocFeatures;
+	size_t m_numAllocFeatures;
 
-	void resize(unsigned int newSize, bool bKeepFeatures= true);
-	int getFeaturePos(unsigned int index);
+	void resize(size_t newSize, bool bKeepFeatures= true);
+	size_t getFeaturePos(size_t index);
 protected:
 	OverwriteMode m_overwriteMode;
 public:
 	Feature* m_pFeatures;
-	unsigned int m_numFeatures;
+	size_t m_numFeatures;
 
 	FeatureList(const char* pName,OverwriteMode overwriteMode=OverwriteMode::Add);
 	virtual ~FeatureList();
@@ -37,21 +37,21 @@ public:
 	const char* getName();
 	void clear();
 	void mult(double factor);
-	double getFactor(unsigned int index) const;
+	double getFactor(size_t index) const;
 	double innerProduct(const FeatureList *inList);
 	void copyMult(double factor,const FeatureList *inList);
 	void addFeatureList(const FeatureList *inList,double factor= 1.0);
-	void add(unsigned int index, double value);
+	void add(size_t index, double value);
 
 	//spawn: all features (indices and values) are spawned by those in inList
 	//[2,3].spawn([1,2,3]) => [2*indexOffset + 1, 2*indexOffset + 2, 2*indexOffset+3
 	//                        , 3*indexOffset+1, 3*indexOffset+2, 3*indexOffset+3]
-	void spawn(const FeatureList *inList, unsigned int indexOffset);
+	void spawn(const FeatureList *inList, size_t indexOffset);
 
 	//adds an offset to all feature indices in a feature list.
 	void offsetIndices(int offset);
 
-	void split(FeatureList *outList1, FeatureList *outList2, unsigned int splitOffset) const;
+	void split(FeatureList *outList1, FeatureList *outList2, size_t splitOffset) const;
 
 	//multiplies all indices by a factor
 	void multIndices(int mult);
